# public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"ShortLinkInterface": {
				"properties": {
					"uuid": {
						"type": "string",
						"description": "Stringified UUIDv4.\nSee [RFC 4112](https://tools.ietf.org/html/rfc4122)",
						"example": "52907745-7672-470e-a803-a2f8feb52944",
						"pattern": "[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-4[0-9A-Fa-f]{3}-[89ABab][0-9A-Fa-f]{3}-[0-9A-Fa-f]{12}"
					},
					"full": {
						"type": "string"
					},
					"short": {
						"type": "string"
					},
					"clicks": {
						"type": "number",
						"format": "double"
					},
					"lastClickAt": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"full"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "url-shortener",
		"version": "1.0.0",
		"description": "Service responsible for the URL shortener",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Paulo Andrade"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/{shortUrl}": {
			"get": {
				"operationId": "FindByShort",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ShortLinkInterface"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "shortUrl",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "Delete",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/ShortLinkInterface"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "shortUrl",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}